Moved optionstate class into project file
