+++ nw/gui/tools/__init__.py
@@ -1,9 +1,7 @@
 # -*- coding: utf-8 -*-
 
 from nw.gui.tools.dochighlight import GuiDocHighlighter
-from nw.gui.tools.optionstate import OptionState
 
 __all__ = [
     "GuiDocHighlighter",
-    "OptionState",
 ]
